+++ code/isa/include/utils.h
@@ -9,20 +9,20 @@ using std::vector;
 
 #define PI 3.141592653589793
 
-Array<double, 1, Dynamic> logsumexp(const ArrayXXd& array);
-Array<double, 1, Dynamic> logmeanexp(const ArrayXXd& array);
+Array<float, 1, Dynamic> logsumexp(const ArrayXXf& array);
+Array<float, 1, Dynamic> logmeanexp(const ArrayXXf& array);
 
-ArrayXXd sampleNormal(int m = 1, int n = 1);
-ArrayXXd sampleGamma(int m = 1, int n = 1, int k = 1);
+ArrayXXf sampleNormal(int m = 1, int n = 1);
+ArrayXXf sampleGamma(int m = 1, int n = 1, int k = 1);
 
-VectorXi argsort(const VectorXd& data);
-MatrixXd covariance(const MatrixXd& data);
-MatrixXd corrcoef(const MatrixXd& data);
-MatrixXd normalize(const MatrixXd& matrix);
+VectorXi argsort(const VectorXf& data);
+MatrixXf covariance(const MatrixXf& data);
+MatrixXf corrcoef(const MatrixXf& data);
+MatrixXf normalize(const MatrixXf& matrix);
 
-double logDetPD(const MatrixXd& matrix);
+float logDetPD(const MatrixXf& matrix);
 
-MatrixXd deleteRows(const MatrixXd& matrix, vector<int> indices);
-MatrixXd deleteCols(const MatrixXd& matrix, vector<int> indices);
+MatrixXf deleteRows(const MatrixXf& matrix, vector<int> indices);
+MatrixXf deleteCols(const MatrixXf& matrix, vector<int> indices);
 
 #endif
