+++ code/isa/src/distribution.cpp
@@ -8,6 +8,6 @@ Distribution::~Distribution() {
 
 
 
-double Distribution::evaluate(const MatrixXd& data) {
+float Distribution::evaluate(const MatrixXf& data) {
 	return -logLikelihood(data).mean() / log(2.) / dim();
 }
