+++ code/isa/src/gsm.cpp
@@ -8,30 +8,30 @@ using std::log;
 using std::rand;
 
 GSM::GSM(int dim, int numScales) : mDim(dim), mNumScales(numScales) {
-	mScales = 1. + ArrayXd::Random(mNumScales) / 4.;
+	mScales = 1. + ArrayXf::Random(mNumScales) / 4.;
 	mScales /= mScales.mean();
 }
 
 
 
-bool GSM::train(const MatrixXd& data, int maxIter, double tol) {
+bool GSM::train(const MatrixXf& data, int maxIter, double tol) {
 	if(data.rows() != mDim)
 		throw Exception("Data has wrong dimensionality.");
 
-	RowVectorXd sqNorms = data.colwise().squaredNorm();
+	RowVectorXf sqNorms = data.colwise().squaredNorm();
 
-	double logLik = logLikelihood(data, sqNorms).mean();
+	float logLik = logLikelihood(data, sqNorms).mean();
 
 	for(int i = 0; i < maxIter; ++i) {
 		// compute unnormalized posterior over mixture components (E)
-		ArrayXXd post = posterior(data, sqNorms);
+		ArrayXXf post = posterior(data, sqNorms);
 
 		// update parameters (M)
 		mScales = (((post.rowwise() * sqNorms.array()).rowwise().mean() + 1e-9)
 			/ (mDim * post.rowwise().mean() + 3e-9)).sqrt();
 
 		if(tol > 0. && i % 5 == 0) {
-			double logLikNew = logLikelihood(data, sqNorms).mean();
+			float logLikNew = logLikelihood(data, sqNorms).mean();
 
 			// check for convergence
 			if(logLikNew - logLik < tol)
@@ -46,8 +46,8 @@ bool GSM::train(const MatrixXd& data, int maxIter, double tol) {
 
 
 
-MatrixXd GSM::sample(int numSamples) {
-	Array<double, 1, Dynamic> scales(1, numSamples);
+MatrixXf GSM::sample(int numSamples) {
+	Array<float, 1, Dynamic> scales(1, numSamples);
 
 	// pick random standard deviations
 	for(int i = 0; i < numSamples; ++i)
@@ -59,14 +59,14 @@ MatrixXd GSM::sample(int numSamples) {
 
 
 
-Array<double, 1, Dynamic> GSM::samplePosterior(const MatrixXd& data) {
-	Array<double, 1, Dynamic> scales(data.cols());
-	ArrayXXd post = posterior(data);
+Array<float, 1, Dynamic> GSM::samplePosterior(const MatrixXf& data) {
+	Array<float, 1, Dynamic> scales(data.cols());
+	ArrayXXf post = posterior(data);
 
 	for(int j = 0; j < post.cols(); ++j) {
 		int i = 0;
-		double urand = static_cast<double>(rand()) / (static_cast<long>(RAND_MAX) + 1l);
-		double cdf;
+		float urand = static_cast<float>(rand()) / (static_cast<long>(RAND_MAX) + 1l);
+		float cdf;
 
 		// compute index
 		for(cdf = post(0, j); cdf < urand; cdf += post(i, j))
@@ -80,15 +80,15 @@ Array<double, 1, Dynamic> GSM::samplePosterior(const MatrixXd& data) {
 
 
 
-ArrayXXd GSM::posterior(const MatrixXd& data) {
+ArrayXXf GSM::posterior(const MatrixXf& data) {
 	return posterior(data, data.colwise().squaredNorm());
 }
 
 
 
-ArrayXXd GSM::posterior(const MatrixXd& data, const RowVectorXd& sqNorms) {
+ArrayXXf GSM::posterior(const MatrixXf& data, const RowVectorXf& sqNorms) {
 	// compute unnormalized log-posterior
-	ArrayXXd posterior = logJoint(data, sqNorms);
+	ArrayXXf posterior = logJoint(data, sqNorms);
 
 	// normalize posterior in a numerically stable way
 	posterior.rowwise() -= posterior.colwise().maxCoeff().eval();
@@ -100,43 +100,43 @@ ArrayXXd GSM::posterior(const MatrixXd& data, const RowVectorXd& sqNorms) {
 
 
 
-ArrayXXd GSM::logJoint(const MatrixXd& data) {
+ArrayXXf GSM::logJoint(const MatrixXf& data) {
 	return logJoint(data, data.colwise().squaredNorm());
 }
 
 
 
-ArrayXXd GSM::logJoint(const MatrixXd& data, const RowVectorXd& sqNorms) {
+ArrayXXf GSM::logJoint(const MatrixXf& data, const RowVectorXf& sqNorms) {
 	return (-0.5 * mScales.square().inverse().matrix() * sqNorms).colwise()
 		- mDim * mScales.log().matrix();
 }
 
 
 
-Array<double, 1, Dynamic> GSM::logLikelihood(const MatrixXd& data) {
+Array<float, 1, Dynamic> GSM::logLikelihood(const MatrixXf& data) {
 	return -energy(data).array() - mDim / 2. * log(2. * PI);
 }
 
 
 
-Array<double, 1, Dynamic> GSM::logLikelihood(const MatrixXd& data, const RowVectorXd& sqNorms) {
+Array<float, 1, Dynamic> GSM::logLikelihood(const MatrixXf& data, const RowVectorXf& sqNorms) {
 	return -energy(data, sqNorms).array() - mDim / 2. * log(2. * PI);
 }
 
 
 
-Array<double, 1, Dynamic> GSM::energy(const MatrixXd& data) {
+Array<float, 1, Dynamic> GSM::energy(const MatrixXf& data) {
 	return -logmeanexp(logJoint(data));
 }
 
 
 
-Array<double, 1, Dynamic> GSM::energy(const MatrixXd& data, const RowVectorXd& sqNorms) {
+Array<float, 1, Dynamic> GSM::energy(const MatrixXf& data, const RowVectorXf& sqNorms) {
 	return -logmeanexp(logJoint(data, sqNorms));
 }
 
 
 
-ArrayXXd GSM::energyGradient(const MatrixXd& data) {
+ArrayXXf GSM::energyGradient(const MatrixXf& data) {
 	return data.array().rowwise() * (posterior(data).colwise() * mScales.square().inverse()).colwise().sum();
 }
